chore: rename pkcache to cache
